# update OneMKL `gemm_batch` call inside `dpnp.matmul`