… tests/test_mathematical.py