In this PR, `dpnp.matmul` implementation that falls back on `gemm_batch` is updated to improve the performance for some cases. The following table compares timing for different cases. Note that improvement is not expected for all the cases.

| Case # | dpnp (old) | dpnp (new) | NumPy |
| ------- | -------- | --------      | -------- |
| 1         | 376 ms | 371 ms       | 580 ms       |
| 2         | 324 ms | 324 ms       | 541 ms       |
| 3         | 125 ms | 130 ms       | 181 ms       |
| 4         | 193 ms | 142 ms       | 150 ms       |
| 5         | 244 ms | 196 ms       | 285 ms       |
| 6         | 251 ms | 163 ms       | 276 ms       |
| 7         | 494 ms | 150 ms       | 582 ms       |
| 8         | 436 ms | 58 ms         | 535 ms       |
| 9         | 129 ms | 131 ms       | 186 ms       |
| 10       | 169 ms | 141 ms       | 164 ms       |
| 11       | 252 ms | 48 ms         | 293 ms       |
| 12       | 254 ms | 20 ms         | 297 ms       |

```
import  dpnp
a = dpnp.ones((2000, 2000, 4, 4))
OUT = dpnp.empty((2000, 2000, 4, 4))

%timeit dpnp.matmul(a, a) # Case 1
out = OUT; 
%timeit dpnp.matmul(a, a, out=out) # Case 2
b = a[::2, ::2]; 
%timeit dpnp.matmul(b, b)  # Case 3
out = OUT[::2, ::2]; 
%timeit dpnp.matmul(b, b, out=out) # Case 4
b = a[:, ::2]; 
%timeit dpnp.matmul(b, b) # Case 5
out = OUT[:, ::2]; 
%timeit dpnp.matmul(b, b, out=out) # Case 6
a = a.transpose(0, 1, 3, 2); 
%timeit dpnp.matmul(a, a) # Case 7
out = OUT; 
%timeit dpnp.matmul(a, a, out=out) # Case 8
b = a[::2, ::2]; 
%timeit dpnp.matmul(b, b) # Case 9
out = OUT[::2, ::2]; 
%timeit dpnp.matmul(b, b, out=out) # Case 10
b = a[:, ::2]; 
%timeit dpnp.matmul(b, b) # Case 11
out = OUT[:, ::2]; 
%timeit dpnp.matmul(b, b, out=out) # Case 12
```

- [x] Have you provided a meaningful PR description?
- [x] Have you added a test, reproducer or referred to issue with a reproducer?
- [x] Have you tested your changes locally for CPU and GPU devices?
- [x] Have you made sure that new changes do not introduce compiler warnings?
- [x] Have you checked performance impact of proposed changes?
- [ ] If this PR is a work in progress, are you filing the PR as a draft?