@vtavana 2024-05-05
Timing for calling `colum_major::gemm` in different scenarios
* dpnp timing on Intel Xeon is slower than NumPy but it is independent of this PR, and should be addressed separately.

| Case # |    NumPy |   old (Iris Xe) |  new (Iris Xe)|   old (Intel Core) | new (Intel Core) |
|-------------|-------------|--------------|---------------|---------------|------------|
| 1| 1.74 s ± 443 ms| 118 ms ± 22.5 ms| 110 ms ± 6.12 ms| 1.71 s ± 439 ms| 1.65 s ± 438 ms|
| 2| 2 s ± 81.7 ms  | 123 ms ± 9.9 ms | 145 ms ± 26.3 ms| 1.88 s ± 427 ms| 1.84 s ± 359 ms|
| 3| 1.82 s ± 472 ms| 125 ms ± 7.88 ms| 149 ms ± 14.2 ms| 2.01 s ± 106 ms| 1.76 s ± 372 ms|
| 4| 2.31 s ± 363 ms| 135 ms ± 28.6 ms| 121 ms ± 18 ms  | 1.99 s ± 393 ms| 1.79 s ± 375 ms|

| Case # | NumPy            | old (PVC) |  new (PVC)|   old (Intel Xeon) | new (Intel Xeon) |
|-------|--------------------|--------------------|--------------------|--------------------|--------------------|
| 1     | 81.8 ms ± 1.32 ms  | 8.35 ms ± 20.6 µs  | 8.47 ms ± 73.1 µs  | 159 ms ± 7.38 ms   | 155 ms ± 6.39 ms   |
| 2     | 81.4 ms ± 1.14 ms  | 8.8 ms ± 62.2 µs   | 8.73 ms ± 43.2 µs  | 157 ms ± 8.78 ms   | 158 ms ± 5.76 ms   |
| 3     | 83.4 ms ± 1.13 ms  | 8.42 ms ± 67.1 µs  | 8.46 ms ± 64 µs    | 156 ms ± 3.87 ms   | 157 ms ± 5.03 ms   |
| 4     | 83.4 ms ± 1.04 ms  | 8.64 ms ± 2.76 µs  | 8.45 ms ± 78.6 µs  | 153 ms ± 8.33 ms   | 152 ms ± 3.45 ms   |

```
import  dpnp
size = 4096
device="gpu"
a = dpnp.ones((size, size), order="C", device=device)
b = dpnp.ones((size, size), order="C", device=device)
%timeit dpnp.matmul(a, b) # Case 1

a = dpnp.ones((size, size), order="C", device=device)
b = dpnp.ones((size, size), order="F", device=device)
%timeit dpnp.matmul(a, b) # Case 2

a = dpnp.ones((size, size), order="F", device=device)
b = dpnp.ones((size, size), order="C", device=device)
%timeit dpnp.matmul(a, b) # Case 3

a = dpnp.ones((size, size), order="F", device=device)
b = dpnp.ones((size, size), order="F", device=device)
%timeit dpnp.matmul(a, b) # Case 4
```