@vtavana 2024-05-05
Timing for calling `gemm_batch` in different scenarios
* dpnp timing for cases when one input is c-contiguous and the other one is f-contiguous is not better than old version on PVC.

| Case # |    NumPy |   old (Iris Xe) |  new (Iris Xe)|   old (Intel Core) | new (Intel Core) |
|--------|---------------------|---------------------|---------------------|---------------------|---------------------|
| 1      | 3.66 s ± 426 ms     | 1.91 s ± 234 ms     | 1.89 s ± 224 ms     | 903 ms ± 109 ms     | 1.02 s ± 75.6 ms   |
| 2      | 3.31 s ± 392 ms     | 5.65 s ± 3.89 s     | 1.74 s ± 268 ms     | 2.19 s ± 129 ms     | 899 ms ± 336 ms     |
| 3      | 3.4 s ± 481 ms      | 2.81 s ± 341 ms     | 443 ms ± 36.9 ms    | 1.91 s ± 45.1 ms    | 825 ms ± 296 ms     |
| 4      | 3.71 s ± 588 ms     | 4.71 s ± 2.27 s     | 1.97 s ± 264 ms     | 2.94 s ± 356 ms     | 871 ms ± 193 ms     |


| Case # | NumPy            | old (PVC) |  new (PVC)|   old (Intel Xeon) | new (Intel Xeon) |
|-------|---------------------|---------------------|---------------------|---------------------|---------------------|
| 1     | 3.67 s ± 2.25 ms    | 57.1 ms ± 862 µs    | 57.6 ms ± 579 µs    | 586 ms ± 47.6 ms    | 497 ms ± 30.9 ms    |
| 2     | 3.78 s ± 1.69 ms    | 68.6 ms ± 471 µs    | 116 ms ± 588 µs     | 850 ms ± 36.1 ms    | 488 ms ± 16.1 ms    |
| 3     | 3.56 s ± 329 µs     | 68.6 ms ± 477 µs    | 75.3 ms ± 566 µs    | 869 ms ± 42.7 ms    | 510 ms ± 30.2 ms    |
| 4     | 3.63 s ± 1.42 ms    | 77 ms ± 38.8 µs     | 58.2 ms ± 589 µs    | 1.06 s ± 28.1 ms    | 472 ms ± 25.2 ms    |

```
import  dpnp
size = 4095
device="gpu"
a = dpnp.ones((size, size, 4, 4), device=device)
b = dpnp.ones((size, size, 4, 4), device=device)
%timeit dpnp.matmul(a, b) # Case 1, both C

a = dpnp.ones((size, size, 4, 4), device=device)
b = dpnp.ones((size, size, 4, 4), device=device)
b = b.transpose(0, 1, 3, 2)
%timeit dpnp.matmul(a, b) # Case 2, a is C, b is F

a = dpnp.ones((size, size, 4, 4), device=device)
b = dpnp.ones((size, size, 4, 4), device=device)
a = a.transpose(0, 1, 3, 2)
%timeit dpnp.matmul(a, b) # Case 3, a is F, b is C

a = dpnp.ones((size, size, 4, 4), device=device)
b = dpnp.ones((size, size, 4, 4), device=device)
a = a.transpose(0, 1, 3, 2)
b = b.transpose(0, 1, 3, 2)
%timeit dpnp.matmul(a, b) # Case 4, both F
```